requirements.txt
Django==4.1
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.3
django-phonenumber-field[phonenumbers]